requirements.txt
aioredis==1.3.1
asgiref==3.4.1
async-timeout==4.0.2
attrs==21.4.0
autobahn==21.2.1
Automat==20.2.0
certifi==2021.5.30
cffi==1.15.1
constantly==15.1.0
cryptography==37.0.4
daphne==3.0.2
Django==3.2.13
django-admin-inline-paginator==0.3.0
django-cors-headers==3.10.0
django-environ==0.10.0
django-model-utils==4.2.0
django-notifications-hq==1.7.0
django-taggit==3.1.0
djangorestframework==3.10.1
hiredis==2.0.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
jsonfield==3.1.0
msgpack==1.0.4
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyOpenSSL==22.0.0
pytz==2022.1
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
swapper==1.3.0
Twisted==22.4.0
txaio==22.2.1
typing_extensions==4.1.0
websocket-client==1.3.1
zope.interface==5.4.0
reportlab==3.6.6
